Grpc - macros moved to client/server modules (not reason why to keep them in common module anymore)
