+++ grpc-client/src/main/scala/com/avast/cactus/grpc/client/client.scala
@@ -5,14 +5,15 @@ import java.util.concurrent.Executor
 import io.grpc._
 import io.grpc.stub.AbstractStub
 
-import scala.concurrent.{ExecutionContext, Future}
+import scala.concurrent.ExecutionContext
 import scala.language.experimental.macros
 
 package object client {
 
   implicit class MapClient(val channel: Channel) extends AnyVal {
-    def createMappedClient[GrpcClientStub <: AbstractStub[GrpcClientStub], MT](interceptors: ClientAsyncInterceptor*)(implicit ec: ExecutionContext, ex: Executor): MT =
-      macro ClientMacros.mapClientToTraitWithInterceptors[GrpcClientStub, MT]
+    def createMappedClient[GrpcClientStub <: AbstractStub[GrpcClientStub], MT](
+        interceptors: ClientAsyncInterceptor*)(implicit ec: ExecutionContext, ex: Executor): MT =
+      macro com.avast.cactus.grpc.client.ClientMacros.mapClientToTraitWithInterceptors[GrpcClientStub, MT]
   }
 
 }
